Update types, keys, and cleanup
