+++ src/form-builder/types.ts
@@ -10,12 +10,20 @@ export type FieldProps<V> = Omit<FieldHookConfig<V>, 'required'> & {
   required?: boolean | string;
 };
 
-export * from './types';
 export type CheckboxProps = FieldProps<string> & {
   checked?: boolean;
   content?: string;
   onValueChange?: any;
   value?: boolean;
+};
+
+export type RadioItemProps = {
+  'aria-describedby': string;
+  checked: boolean;
+  label: string;
+  name: string;
+  value: string;
+  radioOptionSelected: () => void;
 };
 
 export type CheckboxGroupProps = FieldProps<string> & {
